Resources Page
- Rename starter-packs to resources
- Add a resources route
- Update resources to match hackuci's
