+++ src/app/views/resources/Resources.scss
