+++ src/assets/data/resources-info.js
@@ -1,0 +1,118 @@
+let starterPackData = [
+  {
+    name: "User Interface",
+    description: "Build a frontend you can interact with!",
+    packs: [
+      {
+        name: "React",
+        link: "https://beta.reactjs.org/learn",
+        tooltip:
+          "Learn how to use React, a JavaScript library for building user interfaces"
+      },
+      {
+        name: "ZotHacks React Frontend",
+        link: "https://github.com/hackuci/zothacks-2020-frontend",
+        tooltip:
+          "Checkout the frontend project and video tutorial from ZotHacks 2020! A great first step to getting started with a React frontend."
+      },
+      {
+        name: "Figma Giveaway",
+        link: "https://docs.google.com/forms/d/e/1FAIpQLSfeaWeXJvj_ZbZZCCb54gh83RS4hmz30g4H_L0jPKqP7KwFHQ/viewform",
+        tooltip: 
+          "Figma is giving away $100 worth of Figma swag to anyone who verifies their education status over the weekend!"
+      },
+      {
+        name: "iOS App Development Workshop",
+        link: "https://www.youtube.com/watch?v=50Uxl6hz3Ug",
+        tooltip:
+          "CodePath at UCI walks you through the basics of Xcode and how to build a simple iOS app for displaying a list of items."
+      },
+      {
+        name: "Android Zothacks Workshop",
+        link: "https://www.youtube.com/watch?v=u0tNBjglS2M",
+        tooltip:
+          "CodePath at UCI teaches you how to build a simple To-Do app in Android Studio."
+      }
+    ]
+  },
+  {
+    name: "Backend",
+    description: "Build an API for data management and processing!",
+    packs: [
+      {
+        name: "ZotHacks Flask Backend",
+        link: "https://github.com/hackuci/zothacks-2020-backend",
+        tooltip:
+          "Checkout the backend project and video tutorial from ZotHacks 2020! A great first step to getting started with a Flask backend."
+      },
+      {
+        name: "Free Public APIs",
+        link: "https://github.com/public-apis/public-apis",
+        tooltip:
+          "A collective list of free APIs for use in software and web development."
+      },
+      {
+        name: "PeterPortal - An Unofficial API for UCI",
+        link: "https://api.peterportal.org/",
+        tooltip:
+          "PeterPortal API provides consolidated data about UCI professors, courses, and grades."
+      },
+    ]
+  },
+  {
+    name: "Gaming",
+    description: "Build the next chart topping indie game!",
+    packs: [
+      {
+        name: "Unity",
+        link: "https://docs.unity3d.com/Manual/index.html",
+        tooltip:
+          "Official documentation for Unity, a cross-platform game engine"
+      },
+      {
+        name: "Godot",
+        link: "https://godotengine.org/",
+        tooltip:
+          "Godot provides a huge set of common tools for game development"
+      },
+      {
+        name: "PyGame",
+        link: "https://www.pygame.org/wiki/GettingStarted",
+        tooltip:
+          "Official documentation for PyGame, a cross-platform set of Python modules designed for writing video games"
+      }
+    ]
+  },
+  {
+    name: "Tools",
+    description: "Gotta build your projects with something!",
+    packs: [
+      {
+        name: "Git",
+        link: "https://github.com/zpinto/learn-git-with-hack",
+        tooltip:
+          "A complete tutorial on how to get started with Git, a version-control system"
+      },
+      {
+        name: "VS Code",
+        link: "https://code.visualstudio.com/",
+        tooltip: 
+          "A streamlined code editor with support for development operations like debugging, task running, and version control."
+      },
+      {
+        name: "Ngrok",
+        link: "https://ngrok.io",
+        tooltip:
+          "Instantly create a public HTTPS url for a website running locally on your development machine using Ngrok"
+      },
+      {
+        name: "Postman",
+        link: "https://www.getpostman.com/",
+        tooltip:
+          "A great tool for testing RESTful APIs made by others or yourself"
+      }
+    ]
+  }
+];
+
+export { starterPackData };
